+++ src/Silverback.Core.Model/Messaging/Publishing/CommandPublisherExtensions.cs
@@ -2,6 +2,7 @@
 // This code is licensed under MIT license (see LICENSE file for details)
 
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using Silverback.Messaging.Messages;
 using Silverback.Util;
@@ -9,7 +10,7 @@ using Silverback.Util;
 namespace Silverback.Messaging.Publishing;
 
 /// <summary>
-///     Adds the <see cref="ExecuteCommand" /> and <see cref="ExecuteCommandAsync" /> methods to the <see cref="IPublisher" /> interface.
+///     Adds the <see cref="ExecuteCommand" /> and <see cref="ExecuteCommandAsync(IPublisher,ICommand,CancellationToken)" /> methods to the <see cref="IPublisher" /> interface.
 /// </summary>
 public static class CommandPublisherExtensions
 {
@@ -64,14 +65,72 @@ public static class CommandPublisherExtensions
     /// <param name="commandMessage">
     ///     The command to be executed.
     /// </param>
-    /// <param name="throwIfUnhandled">
-    ///     A boolean value indicating whether an exception must be thrown if no subscriber is handling the message.
+    /// <param name="cancellationToken">
+    ///     The <see cref="CancellationToken" /> that can be used to cancel the operation.
     /// </param>
     /// <returns>
     ///     A <see cref="Task" /> representing the asynchronous operation.
     /// </returns>
-    public static Task ExecuteCommandAsync(this IPublisher publisher, ICommand commandMessage, bool throwIfUnhandled = true) =>
-        Check.NotNull(publisher, nameof(publisher)).PublishAsync(commandMessage, throwIfUnhandled);
+    public static Task ExecuteCommandAsync(
+        this IPublisher publisher,
+        ICommand commandMessage,
+        CancellationToken cancellationToken = default) =>
+        Check.NotNull(publisher, nameof(publisher)).PublishAsync(commandMessage, true, cancellationToken);
+
+    /// <summary>
+    ///     Executes the specified command publishing it to the internal bus. The message will be forwarded to its subscribers and the
+    ///     method will not complete until all subscribers have processed it (unless using Silverback.Integration to produce and consume
+    ///     the message through a message broker).
+    /// </summary>
+    /// <param name="publisher">
+    ///     The <see cref="IPublisher" />.
+    /// </param>
+    /// <param name="commandMessage">
+    ///     The command to be executed.
+    /// </param>
+    /// <param name="throwIfUnhandled">
+    ///     A boolean value indicating whether an exception must be thrown if no subscriber is handling the message.
+    /// </param>
+    /// <param name="cancellationToken">
+    ///     The <see cref="CancellationToken" /> that can be used to cancel the operation.
+    /// </param>
+    /// <returns>
+    ///     A <see cref="Task" /> representing the asynchronous operation.
+    /// </returns>
+    public static Task ExecuteCommandAsync(
+        this IPublisher publisher,
+        ICommand commandMessage,
+        bool throwIfUnhandled,
+        CancellationToken cancellationToken = default) =>
+        Check.NotNull(publisher, nameof(publisher)).PublishAsync(commandMessage, throwIfUnhandled, cancellationToken);
+
+    /// <summary>
+    ///     Executes the specified command publishing it to the internal bus. The message will be forwarded to its subscribers and the
+    ///     method will not complete until all subscribers have processed it (unless using Silverback.Integration to produce and consume
+    ///     the message through a message broker).
+    /// </summary>
+    /// <typeparam name="TResult">
+    ///     The type of the result that is expected to be returned by the subscribers.
+    /// </typeparam>
+    /// <param name="publisher">
+    ///     The <see cref="IPublisher" />.
+    /// </param>
+    /// <param name="commandMessage">
+    ///     The command to be executed.
+    /// </param>
+    /// <param name="cancellationToken">
+    ///     The <see cref="CancellationToken" /> that can be used to cancel the operation.
+    /// </param>
+    /// <returns>
+    ///     A <see cref="Task{TResult}" /> representing the asynchronous operation. The Task result contains the
+    ///     command result.
+    /// </returns>
+    public static async Task<TResult> ExecuteCommandAsync<TResult>(
+        this IPublisher publisher,
+        ICommand<TResult> commandMessage,
+        CancellationToken cancellationToken = default) =>
+        (await Check.NotNull(publisher, nameof(publisher)).PublishAsync<TResult>(commandMessage, true, cancellationToken).ConfigureAwait(false))
+        .Single();
 
     /// <summary>
     ///     Executes the specified command publishing it to the internal bus. The message will be forwarded to its subscribers and the
@@ -90,6 +149,9 @@ public static class CommandPublisherExtensions
     /// <param name="throwIfUnhandled">
     ///     A boolean value indicating whether an exception must be thrown if no subscriber is handling the message.
     /// </param>
+    /// <param name="cancellationToken">
+    ///     The <see cref="CancellationToken" /> that can be used to cancel the operation.
+    /// </param>
     /// <returns>
     ///     A <see cref="Task{TResult}" /> representing the asynchronous operation. The Task result contains the
     ///     command result.
@@ -97,6 +159,8 @@ public static class CommandPublisherExtensions
     public static async Task<TResult> ExecuteCommandAsync<TResult>(
         this IPublisher publisher,
         ICommand<TResult> commandMessage,
-        bool throwIfUnhandled = true) =>
-        (await Check.NotNull(publisher, nameof(publisher)).PublishAsync<TResult>(commandMessage, throwIfUnhandled).ConfigureAwait(false)).Single();
+        bool throwIfUnhandled,
+        CancellationToken cancellationToken = default) =>
+        (await Check.NotNull(publisher, nameof(publisher)).PublishAsync<TResult>(commandMessage, throwIfUnhandled, cancellationToken).ConfigureAwait(false))
+        .Single();
 }