+++ src/Silverback.Core/Messaging/Publishing/IStreamPublisher.cs
@@ -2,6 +2,7 @@
 // This code is licensed under MIT license (see LICENSE file for details)
 
 using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 using Silverback.Messaging.Messages;
 
@@ -36,9 +37,12 @@ internal interface IStreamPublisher
     /// <param name="streamProvider">
     ///     The <see cref="IMessageStreamProvider" /> to be used to generate the streams to be published.
     /// </param>
+    /// <param name="cancellationToken">
+    ///     The <see cref="CancellationToken" /> that can be used to cancel the operation.
+    /// </param>
     /// <returns>
     ///     A <see cref="Task{TResult}" /> representing the asynchronous operation. The task result contains a
     ///     collection of <see cref="Task" /> that will complete when each subscriber completes.
     /// </returns>
-    Task<IReadOnlyCollection<Task>> PublishAsync(IMessageStreamProvider streamProvider);
+    Task<IReadOnlyCollection<Task>> PublishAsync(IMessageStreamProvider streamProvider, CancellationToken cancellationToken = default);
 }