+++ src/Silverback.Core/Messaging/Publishing/StreamPublisher.cs
@@ -2,6 +2,7 @@
 // This code is licensed under MIT license (see LICENSE file for details)
 
 using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 using Silverback.Messaging.Messages;
 
@@ -19,6 +20,6 @@ internal sealed class StreamPublisher : IStreamPublisher
     public IReadOnlyCollection<Task> Publish(IMessageStreamProvider streamProvider) =>
         _publisher.Publish<Task>(streamProvider);
 
-    public async Task<IReadOnlyCollection<Task>> PublishAsync(IMessageStreamProvider streamProvider) =>
-        await _publisher.PublishAsync<Task>(streamProvider).ConfigureAwait(false);
+    public Task<IReadOnlyCollection<Task>> PublishAsync(IMessageStreamProvider streamProvider, CancellationToken cancellationToken = default) =>
+        _publisher.PublishAsync<Task>(streamProvider, cancellationToken);
 }