+++ tests/Silverback.Core.Tests/Messaging/Publishing/PublisherFixture.Behaviors.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Globalization;
+using System.Threading;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
@@ -122,7 +123,7 @@ public partial class PublisherFixture
             _changedMessageFactory = changedMessageFactory;
         }
 
-        public ValueTask<IReadOnlyCollection<object?>> HandleAsync(object message, MessageHandler next) =>
+        public ValueTask<IReadOnlyCollection<object?>> HandleAsync(object message, MessageHandler next, CancellationToken cancellationToken) =>
             next(message is TSourceType ? _changedMessageFactory(message) : message);
     }
 
@@ -139,7 +140,7 @@ public partial class PublisherFixture
 
         public int ExitCount { get; private set; }
 
-        public ValueTask<IReadOnlyCollection<object?>> HandleAsync(object message, MessageHandler next)
+        public ValueTask<IReadOnlyCollection<object?>> HandleAsync(object message, MessageHandler next, CancellationToken cancellationToken)
         {
             _calls?.Add("unsorted");
 
@@ -169,7 +170,7 @@ public partial class PublisherFixture
 
         public int ExitCount { get; private set; }
 
-        public ValueTask<IReadOnlyCollection<object?>> HandleAsync(object message, MessageHandler next)
+        public ValueTask<IReadOnlyCollection<object?>> HandleAsync(object message, MessageHandler next, CancellationToken cancellationToken)
         {
             _calls?.Add(SortIndex.ToString(CultureInfo.InvariantCulture));