+++ tests/Silverback.Integration.Tests/Messaging/Producing/Routing/OutboundRouterBehaviorFixture.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using FluentAssertions;
 using NSubstitute;
@@ -40,7 +41,7 @@ public class OutboundRouterBehaviorFixture
         AddProducer<TestEventTwo>("topic2");
         int nextBehaviorCalls = 0;
 
-        await _behavior.HandleAsync(new TestEventOne(), _ => Next(ref nextBehaviorCalls));
+        await _behavior.HandleAsync(new TestEventOne(), _ => Next(ref nextBehaviorCalls), CancellationToken.None);
 
         nextBehaviorCalls.Should().Be(1);
         _messageWrapper.ReceivedCalls().Should().BeEmpty();
@@ -54,8 +55,8 @@ public class OutboundRouterBehaviorFixture
         TestEventOne message1 = new();
         TestEventTwo message2 = new();
 
-        await _behavior.HandleAsync(message1, _ => Next());
-        await _behavior.HandleAsync(message2, _ => Next());
+        await _behavior.HandleAsync(message1, _ => Next(), CancellationToken.None);
+        await _behavior.HandleAsync(message2, _ => Next(), CancellationToken.None);
 
         _messageWrapper.ReceivedCalls().Count().Should().Be(2);
         await _messageWrapper.Received(1).WrapAndProduceAsync(message1, Arg.Any<IPublisher>(), ArgIsProducers(producer1));
@@ -69,7 +70,7 @@ public class OutboundRouterBehaviorFixture
         IProducer producer2 = AddProducer<TestEventOne>("topic2");
         TestEventOne message = new();
 
-        await _behavior.HandleAsync(message, _ => Next());
+        await _behavior.HandleAsync(message, _ => Next(), CancellationToken.None);
 
         _messageWrapper.ReceivedCalls().Count().Should().Be(1);
         await _messageWrapper.Received(1).WrapAndProduceAsync(message, Arg.Any<IPublisher>(), ArgIsProducers(producer1, producer2));
@@ -83,8 +84,8 @@ public class OutboundRouterBehaviorFixture
         List<TestEventOne> messages1 = [new(), new()];
         TestEventTwo[] messages2 = [new(), new()];
 
-        await _behavior.HandleAsync(messages1, _ => Next());
-        await _behavior.HandleAsync(messages2, _ => Next());
+        await _behavior.HandleAsync(messages1, _ => Next(), CancellationToken.None);
+        await _behavior.HandleAsync(messages2, _ => Next(), CancellationToken.None);
 
         _messageWrapper.ReceivedCalls().Count().Should().Be(2);
         await _messageWrapper.Received(1).WrapAndProduceBatchAsync(messages1, Arg.Any<IPublisher>(), ArgIsProducers(producer1));
@@ -99,8 +100,8 @@ public class OutboundRouterBehaviorFixture
         IEnumerable<TestEventOne> messages1 = new TestEventOne[] { new(), new() }.ToPureEnumerable();
         IEnumerable<TestEventTwo> messages2 = new TestEventTwo[] { new(), new() }.ToPureEnumerable();
 
-        await _behavior.HandleAsync(messages1, _ => Next());
-        await _behavior.HandleAsync(messages2, _ => Next());
+        await _behavior.HandleAsync(messages1, _ => Next(), CancellationToken.None);
+        await _behavior.HandleAsync(messages2, _ => Next(), CancellationToken.None);
 
         _messageWrapper.ReceivedCalls().Count().Should().Be(2);
         await _messageWrapper.Received(1).WrapAndProduceBatchAsync(messages1, Arg.Any<IPublisher>(), ArgIsProducers(producer1));
@@ -115,8 +116,8 @@ public class OutboundRouterBehaviorFixture
         IAsyncEnumerable<TestEventOne> messages1 = new TestEventOne[] { new(), new() }.ToAsyncEnumerable();
         IAsyncEnumerable<TestEventTwo> messages2 = new TestEventTwo[] { new(), new() }.ToAsyncEnumerable();
 
-        await _behavior.HandleAsync(messages1, _ => Next());
-        await _behavior.HandleAsync(messages2, _ => Next());
+        await _behavior.HandleAsync(messages1, _ => Next(), CancellationToken.None);
+        await _behavior.HandleAsync(messages2, _ => Next(), CancellationToken.None);
 
         _messageWrapper.ReceivedCalls().Count().Should().Be(2);
         await _messageWrapper.Received(1).WrapAndProduceBatchAsync(messages1, Arg.Any<IPublisher>(), ArgIsProducers(producer1));
@@ -131,8 +132,8 @@ public class OutboundRouterBehaviorFixture
         Tombstone<TestEventOne> message1 = new("1");
         Tombstone<TestEventTwo> message2 = new("2");
 
-        await _behavior.HandleAsync(message1, _ => Next());
-        await _behavior.HandleAsync(message2, _ => Next());
+        await _behavior.HandleAsync(message1, _ => Next(), CancellationToken.None);
+        await _behavior.HandleAsync(message2, _ => Next(), CancellationToken.None);
 
         _messageWrapper.ReceivedCalls().Count().Should().Be(2);
         await _messageWrapper.Received(1).WrapAndProduceAsync(message1, Arg.Any<IPublisher>(), ArgIsProducers(producer1));
@@ -147,8 +148,8 @@ public class OutboundRouterBehaviorFixture
         List<Tombstone<TestEventOne>> messages1 = [new("1"), new("2")];
         Tombstone<TestEventTwo>[] messages2 = [new("1"), new("2")];
 
-        await _behavior.HandleAsync(messages1, _ => Next());
-        await _behavior.HandleAsync(messages2, _ => Next());
+        await _behavior.HandleAsync(messages1, _ => Next(), CancellationToken.None);
+        await _behavior.HandleAsync(messages2, _ => Next(), CancellationToken.None);
 
         _messageWrapper.ReceivedCalls().Count().Should().Be(2);
         await _messageWrapper.Received(1).WrapAndProduceBatchAsync(messages1, Arg.Any<IPublisher>(), ArgIsProducers(producer1));